scraper/requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
lxml
Pillow
python-dotenv
pydantic
